==> Rscripts/00.CreateNecessaryDirectories.R <==
dir.create("Data/unzipped/")
dir.create("Output/")
dir.create("Bashscripts/")
dir.create("Output/AAchanges")
dir.create("Output/bam")
dir.create("Output/bam2")
dir.create("Output/Bash")
dir.create("Output/clean.fq")
dir.create("Output/CSV")
dir.create("Output/Consensus")
dir.create("Output/Fig.1")
dir.create("Output/HammingDistanceFiltering")
dir.create("Output/Mut.freq")
dir.create("Output/MutFreq")
dir.create("Output/Overview_filtered")
dir.create("Output/Overview_output")
dir.create("Output/Overview_outpu_Ref")
dir.create("Output/Ref")
dir.create("Output/sam")
dir.create("Output/sam2")
dir.create("Output/SelCoeff")
dir.create("Output/SeqData")
dir.create("Output/SummaryFig.Filtered")
dir.create("Output/SummaryFigures")
dir.create("Output/SummaryStats")
dir.create("Output/MutFreq/filtered")
