==> Rscripts/12.5.Estimate.mutation.rates.R <==
library(tidyverse)
library(zoo)
library(ggplot2)
library(reshape)
library(ggpubr)
library(ggthemes)
library(sfsmisc)
library(colorspace)
source("Rscripts/label_scientific.R")
source("Rscripts/baseRscript.R")

colors2<-qualitative_hcl(6, palette="Dark3")
col2_light<-qualitative_hcl(6, palette="Set3")
div.colors<-c(colors2[1],col2_light[1],colors2[3],col2_light[3],colors2[5],col2_light[5])
color.genes<-qualitative_hcl(11, palette="Dark3")


TS<-read.csv("Output1A/MutFreq.filtered/Filtered.Ts.Q35.csv",stringsAsFactors = F, row.names = 1)
stops<-TS[TS$Type=="stop",]
#transition rates (only c and g are present)
stopC<-stops[stops$ref =="c",]
stopG<-stops[stops$ref =="g",]

#G to A mutation
mean(stopG$mean) #0.001915474
#C to T mutation
mean(stopC$mean) #0.002165682


Tv1<-read.csv("Output1A/MutFreq.filtered/Filtered.Tv1.MutFreq.Q35.csv",stringsAsFactors = F, row.names = 1)
Tv2<-read.csv("Output1A/MutFreq.filtered/Filtered.Tv2.MutFreq.Q35.csv",stringsAsFactors = F, row.names = 1)

stopTv1<-Tv1[Tv1$Type.tv1=="stop",]
stopTv2<-Tv2[Tv2$Type.tv2=="stop",]

table(stopTv1$ref)
#c   t 
#175  97

table(stopTv2$ref)
#a   c   g   t 
#107  80 138  36 


#C to A mutation
mean(stopTv1$mean[stopTv1$ref=="c"]) #0.0005057763
#T to A mutation
mean(stopTv1$mean[stopTv1$ref=="t"]) #0.0005727744

#A to T mutation
mean(stopTv2$mean[stopTv2$ref=="a"]) #0.0006335722
#C to G mutation
mean(stopTv2$mean[stopTv2$ref=="c"]) # 3.796976e-05
#G to T mutation
mean(stopTv2$mean[stopTv2$ref=="g"]) # 0.001000036
#T to G mutation
mean(stopTv2$mean[stopTv2$ref=="t"]) # 0.0001766808


## mutation rates from Geller
geller<-read.csv("Data/Geller.mutation.rates.csv")
geller<-geller[c(2,3,4,5,7,8,9,10,12,17,18,19),]

Mutrates<-data.frame(mutations=c("CT","GA","CA","TA","AT","CG","GT","TG"),
           Estimated=c(mean(stopG$mean), 
                      mean(stopC$mean),
                      mean(stopTv1$mean[stopTv1$ref=="c"]), 
                      mean(stopTv1$mean[stopTv1$ref=="t"]), 
                      mean(stopTv2$mean[stopTv2$ref=="a"]), 
                      mean(stopTv2$mean[stopTv2$ref=="c"]), 
                      mean(stopTv2$mean[stopTv2$ref=="g"]), 
                      mean(stopTv2$mean[stopTv2$ref=="t"]) ))

geller$mutations<-gsub("U","T",geller$mutations)

Mutrates<-merge(Mutrates, geller[,1:2], by="mutations", all.y = T )
colnames(Mutrates)[2:3]<-c("In vivo estimation", "In vitro")

MutratesM<-melt(Mutrates)
ggplot(MutratesM, aes(x= mutations, y=value, color=variable ))+
        geom_point(size=2)+scale_y_continuous(trans = "log")+
        scale_color_manual(values=colors2[c(1,4)])+
        theme_bw()+
        ylab("Mutation rate")+xlab('Mutation')+
        theme(legend.title = element_blank())
ggsave("Output1A/SummaryFig.Filtered/Estimated.mutation.rates.pdf", width = 6, height = 4)


cor.test(Mutrates$`In vivo estimation`, Mutrates$`In vitro`, method="spearman")
#S = 18, p-value = 0.02793
#rho 0.7857143
